Add files via upload
I uploaded the class_Ib. file, but when I open it, it still opens in the old project. Please make sure to switch the active project to the class_Ib.R project so that it opens separately and shows only that project. I’m not sure how to separate it myself.